+++ .gitignore
@@ -1,5 +1,20 @@
-build
-dist
+build/
+_build/
+*.o
+*.so
+*.a
+*.py[cod]
 *.egg-info
-*.pyc
-setuptools-*
+dist/
+__pycache__
+.DS_Store
+*.deb
+*.dsc
+*.build
+*.changes
+*.orig.*
+packaging/*tar.xz
+library/debian/
+.coverage
+.pytest_cache
+.tox
