+++ library/CHANGELOG.txt
@@ -1,3 +1,8 @@
+0.0.2
+-----
+
+* Support for multiple display sizes/orientations
+
 0.0.1
 -----
