+++ library/README.rst
@@ -1,5 +1,7 @@
-Python ST7735 160x80
-====================
+Python ST7735
+=============
+
+|Build Status| |Coverage Status| |PyPi Package| |Python Versions|
 
 Python library to control an ST7735 TFT LCD display. Allows simple
 drawing on the display without installing a kernel module.
@@ -34,3 +36,12 @@ Modified from 'Modified from 'Adafruit Python ILI9341' written by Tony
 DiCola for Adafruit Industries.' written by Clement Skau.
 
 MIT license, all text above must be included in any redistribution
+
+.. |Build Status| image:: https://travis-ci.com/pimoroni/st7735-python.svg?branch=master
+   :target: https://travis-ci.com/pimoroni/st7735-python
+.. |Coverage Status| image:: https://coveralls.io/repos/github/pimoroni/st7735-python/badge.svg?branch=master
+   :target: https://coveralls.io/github/pimoroni/st7735-python?branch=master
+.. |PyPi Package| image:: https://img.shields.io/pypi/v/st7735.svg
+   :target: https://pypi.python.org/pypi/st7735
+.. |Python Versions| image:: https://img.shields.io/pypi/pyversions/st7735.svg
+   :target: https://pypi.python.org/pypi/st7735
