+++ library/setup.cfg
@@ -1,0 +1,11 @@
+[flake8]
+exclude =
+	test.py
+	.tox,
+	.eggs,
+	.git,
+	__pycache__,
+	build,
+	dist
+ignore =
+	E501
