+++ library/tox.ini
@@ -1,0 +1,21 @@
+[tox]
+envlist = py{27,35},qa
+skip_missing_interpreters = True
+
+[testenv]
+commands =
+	python setup.py install
+	coverage run -m py.test -v -r wsx
+	coverage report
+deps =
+	mock
+	pytest>=3.1
+	pytest-cov
+ 
+[testenv:qa]
+commands =
+	flake8 --ignore E501
+	rstcheck README.rst
+deps =
+	flake8
+	rstcheck
